+++ database/seeds/UserSeeder.php
@@ -2,17 +2,37 @@
 
 use Illuminate\Database\Seeder;
 
-class UserSeeder extends Seeder {
+class UserSeeder extends Seeder
+{
 	/**
 	 * Run the database seeds.
 	 *
 	 * @return void
 	 */
-	public function run() {
+	public function run()
+	{
 		$role = App\Role::create([
 			'name' => 'admin',
 			'display_name' => 'Administrador',
-			'description' => 'Rol administrador',
+			'description' => 'Administrador',
+		]);
+
+		$role = App\Role::create([
+			'name' => 'rrhh',
+			'display_name' => 'RR.HH.',
+			'description' => 'Gestión de Empleados',
+		]);
+
+		$role = App\Role::create([
+			'name' => 'juridica',
+			'display_name' => 'Juridica',
+			'description' => 'Gestión de Contratos',
+		]);
+
+		$role = App\Role::create([
+			'name' => 'financiera',
+			'display_name' => 'Financiera',
+			'description' => 'Impresión de Planillas',
 		]);
 
 		$user = App\User::create([